add dictionary implementation

+++ core-dictionary/src/main/java/com/oppo/tagbase/dict/AppendableDictionary.java
@@ -9,6 +9,6 @@ public interface AppendableDictionary extends Dictionary {
     /**
      * 向字典中添加一个值
      */
-    int add(byte[] element);
+    long add(byte[] element);
 
 }
